return WP_Error in stats step only when WP_DEBUG is set to true

+++ inc/settings/controllers/leyka-class-init-settings-controller.php
@@ -745,13 +745,13 @@ class Leyka_Init_Wizard_Settings_Controller extends Leyka_Wizard_Settings_Contro
         update_option('leyka_plugin_stats_option_needs_sync', time());
         $stats_option_synch_res = leyka_sync_plugin_stats_option();
 
-        // DO NOT return WP_Error!!!! We should save option and go next step anyway
-        //if(is_wp_error($stats_option_synch_res)) {
-        //    return $stats_option_synch_res;
-        //} else {
+        // DO NOT return WP_Error in production!!!! We should save option and go next step anyway
+        if(is_wp_error($stats_option_synch_res) && defined('WP_DEBUG') && WP_DEBUG) {
+            return $stats_option_synch_res;
+        } else {
             return delete_option('leyka_plugin_stats_option_needs_sync')
                 && update_option('leyka_plugin_stats_option_sync_done', time());
-        //}
+        }
 
     }
